Share db

+++ backend/src/network/client_handler.py
@@ -1,3 +1,4 @@
+import os
 from database import DBManager
 from executer import StorageManager
 from constants import Account, Project
@@ -5,6 +6,7 @@ from network.sessions import SessionManager, Session
 from utils import Logger
 import secrets
 import time 
+from shareDB import ShareDBManager
 
 class ClientHandler():
     def __init__(self, manager, socket) -> None:
@@ -16,12 +18,13 @@ class ClientHandler():
         self.db_manager: DBManager = DBManager.get_instance()
         self.session_manager: SessionManager = SessionManager.get_instance()
         self.storage_manager: StorageManager = StorageManager()
+        self.shareDB_manager: ShareDBManager = ShareDBManager.get_instance()
         
         self.account: Account = None
         self.project: Project = None
         self.session: Session = None
 
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def createUser(props):
             if self.db_manager.get_user_by_email(props['email']) != None:
                 return False
@@ -40,7 +43,7 @@ class ClientHandler():
                 "user": self._formatAccountToJson()
             }
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def loginUser(props):
             ret = self.db_manager.get_user_by_email(props['email'])
             if ret is None: return False
@@ -56,7 +59,7 @@ class ClientHandler():
                 "user": self._formatAccountToJson()
             }
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def autoLogin(props: dict):
             user_id = self.db_manager.get_user_id_by_token(props['token'])
             if user_id is None: return False
@@ -71,13 +74,13 @@ class ClientHandler():
                 "user": self._formatAccountToJson()
             }
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def updateUserCredentials(props):
             self.db_manager.update_user_credentials(self.account.id, props["username"], props["email"])
             self.account = Account(self.account.id, props["username"], props["email"])
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def updateUserPassword(props):
             if self.db_manager.get_user_password(self.account.id) != props["oldPassword"]:
                 return False
@@ -85,7 +88,7 @@ class ClientHandler():
             self.db_manager.update_user_password(self.account.id, props["newPassword"])
             return True
 
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def getProjectListForUser(props):
             id_list = self.db_manager.get_projects_for_user(self.account.id)
             if id_list is None: return {"projects": []}
@@ -99,7 +102,7 @@ class ClientHandler():
             }
             
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def createProject(props):
             project_id = self.db_manager.create_project(self.account.id)
             try: 
@@ -116,7 +119,7 @@ class ClientHandler():
                 
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def updateProjectMetadata(props):
             self.storage_manager.update_metadata(id=props["id"], 
                                                  name=props["name"], 
@@ -125,18 +128,19 @@ class ClientHandler():
             
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def deleteProject(props):
             if self.db_manager.is_user_admin(self.account.id, props["id"]):
                 self.db_manager.delete_project(props["id"])
                 self.storage_manager.delete_project(props["id"])
+                self.shareDB_manager.delete_project(props["id"])
 
             else:
                 self.db_manager.delete_project_for_user(self.account.id, props["id"])
 
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def setCurrentProject(project_id):
             if not self.db_manager.check_user_permission(self.account.id, project_id): return False
             is_admin = self.db_manager.is_user_admin(self.account.id, project_id)
@@ -146,49 +150,52 @@ class ClientHandler():
             project_json["isAdmin"] = is_admin
             return project_json
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def getProjectFilePaths(props):
             return self.storage_manager.get_files(self.project.id)
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def createFile(props):
             return self.storage_manager.create_file(self.project.id, props["name"])
 
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def renameFile(props):
+            self.shareDB_manager.rename_file(self.project.id, props["oldName"], props["newName"])
             return self.storage_manager.edit_file_name(self.project.id, props["oldName"], props["newName"])
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def fetchFile(props):
             return self.storage_manager.get_file(self.project.id, props["name"])
 
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def deleteFile(props):
+            self.shareDB_manager.delete_file(self.project.id, props["name"])
             return self.storage_manager.delete_file(self.project.id, props["name"])
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def clientInHomePage(props):
             if self.session is not None:
                 self.session.leave(self)
                 self.session = None
 
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def joinSession(props):
             self.session = self.session_manager.get_session(self.project)
             self.session.join(self)
             return True
 
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def executerCommand(props):
             self.session.controller.send_input(props["command"])
             return True
         
-        @self.socket.eventHandler
-        def autoSave(props):
-            self.storage_manager.save_file(self.project.id, props["name"], props["data"])
+        @self.socket.event_handler
+        def saveFile(props):
+            self.storage_manager.save_file(self.project.id, props["name"], 
+                                           self.shareDB_manager.read_file(self.project.id, props["name"]))
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def getInvitationRequests(props):
             project_id_list = self.db_manager.get_invites_for_user(self.account.id)
             if project_id_list is None: return {"projects": []}
@@ -196,25 +203,27 @@ class ClientHandler():
                 "projects": [self.storage_manager.get_metadata(id) for id in project_id_list]
             }
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def acceptInvite(props):
+            # Check if the user is already in the project
+            if self.db_manager.check_user_permission(self.account.id, props["id"]): return False
             self.db_manager.delete_invite(self.account.id, props["id"])
             self.db_manager.add_user_to_project(self.account.id, props["id"])
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def declineInvite(props):
             self.db_manager.delete_invite(self.account.id, props["id"])
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def getUsersForProject(props):
             id_list = self.db_manager.get_users_for_project(self.project.id)
             if id_list is None: return {"users": []}
             username_list = [self.db_manager.get_user_by_id(id)[1] for id in id_list if id != self.account.id]
             return username_list
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def removeUserFromProject(props):
             user_id = self.db_manager.get_user_by_username(props["username"])[0]
             self.db_manager.remove_user_from_project(user_id, self.project.id)
@@ -223,7 +232,7 @@ class ClientHandler():
                     user.socket.send("goToHome")
             return True
         
-        @self.socket.eventHandler
+        @self.socket.event_handler
         def sendInvite(props):
             user = self.db_manager.get_user_by_username(props["username"])
             if user is None: return False
